delivery form component, no tabs if no pickup
